% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{delete_license_configuration}
\alias{delete_license_configuration}
\title{Deletes an existing license configuration}
\usage{
delete_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] Unique ID of the configuration object to delete.}
}
\description{
Deletes an existing license configuration. This action fails if the configuration is in use.
}
\section{Accepted Parameters}{
\preformatted{delete_license_configuration(
  LicenseConfigurationArn = "string"
)
}
}

