% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{get_license_configuration}
\alias{get_license_configuration}
\title{Returns a detailed description of a license configuration}
\usage{
get_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] ARN of the license configuration being requested.}
}
\description{
Returns a detailed description of a license configuration.
}
\section{Accepted Parameters}{
\preformatted{get_license_configuration(
  LicenseConfigurationArn = "string"
)
}
}

