% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{get_service_settings}
\alias{get_service_settings}
\title{Gets License Manager settings for a region}
\usage{
get_service_settings()
}
\description{
Gets License Manager settings for a region. Exposes the configured S3 bucket, SNS topic, etc., for inspection.
}
\section{Accepted Parameters}{
\preformatted{get_service_settings()
}
}

