% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{list_license_specifications_for_resource}
\alias{list_license_specifications_for_resource}
\title{Returns the license configuration for a resource}
\usage{
list_license_specifications_for_resource(
  ResourceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{ResourceArn}{[required] ARN of an AMI or Amazon EC2 instance that has an associated license configuration.}

\item{MaxResults}{Maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Returns the license configuration for a resource.
}
\examples{

}
