% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{list_resource_inventory}
\alias{list_resource_inventory}
\title{Returns a detailed list of resources}
\usage{
list_resource_inventory(MaxResults = NULL, NextToken = NULL,
  Filters = NULL)
}
\arguments{
\item{MaxResults}{Maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned \code{NextToken} value.}

\item{NextToken}{Token for the next set of results.}

\item{Filters}{One or more filters.}
}
\description{
Returns a detailed list of resources.
}
\section{Accepted Parameters}{
\preformatted{list_resource_inventory(
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      Name = "string",
      Condition = "EQUALS"|"NOT_EQUALS"|"BEGINS_WITH"|"CONTAINS",
      Value = "string"
    )
  )
)
}
}

