% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists tags attached to a resource}
\usage{
list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] ARN for the resource.}
}
\description{
Lists tags attached to a resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  ResourceArn = "string"
)
}
}

