% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_package.R
\docType{package}
\name{paws.licensemanager-package}
\alias{paws.licensemanager}
\alias{paws.licensemanager-package}
\title{paws.licensemanager: AWS License Manager}
\description{
*This is the AWS License Manager API
    Reference.* It provides descriptions, syntax, and usage examples for
    each of the actions and data types for License Manager. The topic for
    each action shows the Query API request parameters and the XML
    response. You can also view the XML request elements in the WSDL.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
