% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Attach one of more tags to any resource}
\usage{
tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ResourceArn}{[required] Resource of the ARN to be tagged.}

\item{Tags}{[required] Names of the tags to attach to the resource.}
}
\description{
Attach one of more tags to any resource.
}
\examples{

}
