% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Remove tags from a resource}
\usage{
untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
\arguments{
\item{ResourceArn}{[required] ARN of the resource.}

\item{TagKeys}{[required] List keys identifying tags to remove.}
}
\description{
Remove tags from a resource.
}
\examples{

}
