% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{update_license_specifications_for_resource}
\alias{update_license_specifications_for_resource}
\title{Adds or removes license configurations for a specified AWS resource}
\usage{
update_license_specifications_for_resource(ResourceArn,
  AddLicenseSpecifications = NULL, RemoveLicenseSpecifications = NULL)
}
\arguments{
\item{ResourceArn}{[required] ARN for an AWS server resource.}

\item{AddLicenseSpecifications}{License configuration ARNs to be added to a resource.}

\item{RemoveLicenseSpecifications}{License configuration ARNs to be removed from a resource.}
}
\description{
Adds or removes license configurations for a specified AWS resource. This operation currently supports updating the license specifications of AMIs, instances, and hosts. Launch templates and AWS CloudFormation templates are not managed from this operation as those resources send the license configurations directly to a resource creation operation, such as \code{RunInstances}.
}
\section{Accepted Parameters}{
\preformatted{update_license_specifications_for_resource(
  ResourceArn = "string",
  AddLicenseSpecifications = list(
    list(
      LicenseConfigurationArn = "string"
    )
  ),
  RemoveLicenseSpecifications = list(
    list(
      LicenseConfigurationArn = "string"
    )
  )
)
}
}

