% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.licensemanager_operations.R
\name{update_service_settings}
\alias{update_service_settings}
\title{Updates License Manager service settings}
\usage{
update_service_settings(S3BucketArn = NULL, SnsTopicArn = NULL,
  OrganizationConfiguration = NULL,
  EnableCrossAccountsDiscovery = NULL)
}
\arguments{
\item{S3BucketArn}{ARN of the Amazon S3 bucket where License Manager information is stored.}

\item{SnsTopicArn}{ARN of the Amazon SNS topic used for License Manager alerts.}

\item{OrganizationConfiguration}{Integrates AWS Organizations with License Manager for cross-account discovery.}

\item{EnableCrossAccountsDiscovery}{Activates cross-account discovery.}
}
\description{
Updates License Manager service settings.
}
\section{Accepted Parameters}{
\preformatted{update_service_settings(
  S3BucketArn = "string",
  SnsTopicArn = "string",
  OrganizationConfiguration = list(
    EnableIntegration = TRUE|FALSE
  ),
  EnableCrossAccountsDiscovery = TRUE|FALSE
)
}
}

